fix: move nodeJS request methods back to node SDK
