+++ sdk/js-cloud-server/__tests__/request.spec.ts
@@ -5,45 +5,11 @@ global.fetch = fetch
 
 const fetchRequestMock = fetch as jest.MockedFn<typeof fetch>
 
-import { dvcDefaultLogger } from '../src/utils/logger'
-const logger = dvcDefaultLogger()
-import { publishEvents, getEnvironmentConfig, post, get } from '../src/request'
+import { post, get } from '../src/request'
 
 describe('request.ts Unit Tests', () => {
     beforeEach(() => {
         fetchRequestMock.mockReset()
-    })
-
-    describe('publishEvents', () => {
-        it('should throw errors for missing sdkKey / config', async () => {
-            await expect(() => publishEvents(logger, null, [])).rejects.toThrow(
-                'DevCycle is not yet initialized to publish events.',
-            )
-        })
-    })
-
-    describe('getEnvironmentConfig', () => {
-        it('should get environment config', async () => {
-            const url = 'https://test.devcycle.com/config'
-            const etag = 'etag_value'
-            fetchRequestMock.mockResolvedValue(
-                new Response('', { status: 200 }) as any,
-            )
-
-            const res = await getEnvironmentConfig(url, 60000, etag)
-            expect(res.status).toEqual(200)
-            expect(fetchRequestMock).toBeCalledWith(url, {
-                headers: {
-                    'If-None-Match': etag,
-                    'Content-Type': 'application/json',
-                },
-                retries: 1,
-                retryDelay: expect.any(Function),
-                retryOn: expect.any(Function),
-                method: 'GET',
-                signal: expect.any(AbortSignal),
-            })
-        })
     })
 
     describe('get', () => {
