+++ sdk/js-cloud-server/src/request.ts
@@ -1,11 +1,8 @@
-import { SDKEventBatchRequestBody, DVCLogger } from '@devcycle/types'
 import { DVCPopulatedUser } from './models/populatedUser'
 import { DevCycleEvent, DevCycleOptions } from './types'
 import fetchWithRetry, { RequestInitWithRetry } from 'fetch-retry'
 
 export const HOST = '.devcycle.com'
-export const EVENT_URL = 'https://events'
-export const EVENTS_PATH = '/v1/events/batch'
 
 const BUCKETING_BASE = 'https://bucketing-api'
 const VARIABLES_PATH = '/v1/variables'
@@ -37,9 +34,7 @@ const retryOnRequestError: retryOnRequestErrorFunc = (retries) => {
     return (attempt, error, response) => {
         if (attempt >= retries) {
             return false
-        }
-
-        if (response && response?.status < 500) {
+        } else if (response && response?.status < 500) {
             return false
         }
 
@@ -64,46 +59,6 @@ const handleResponse = async (res: Response) => {
     }
 
     return res
-}
-
-export async function publishEvents(
-    logger: DVCLogger,
-    sdkKey: string | null,
-    eventsBatch: SDKEventBatchRequestBody,
-    eventsBaseURLOverride?: string,
-): Promise<Response> {
-    if (!sdkKey) {
-        throw new Error('DevCycle is not yet initialized to publish events.')
-    }
-    const url = eventsBaseURLOverride
-        ? `${eventsBaseURLOverride}${EVENTS_PATH}`
-        : `${EVENT_URL}${HOST}${EVENTS_PATH}`
-    return await post(
-        url,
-        {
-            body: JSON.stringify({ batch: eventsBatch }),
-        },
-        sdkKey,
-    )
-}
-
-export async function getEnvironmentConfig(
-    url: string,
-    requestTimeout: number,
-    etag?: string,
-): Promise<Response> {
-    const headers: Record<string, string> = etag
-        ? { 'If-None-Match': etag }
-        : {}
-
-    return await getWithTimeout(
-        url,
-        {
-            headers: headers,
-            retries: 1,
-        },
-        requestTimeout,
-    )
 }
 
 export async function getAllFeatures(
@@ -228,23 +183,6 @@ export async function get(
     })
 
     return handleResponse(res)
-}
-
-async function getWithTimeout(
-    url: string,
-    requestConfig: RequestInit | RequestInitWithRetry,
-    timeout: number,
-): Promise<Response> {
-    const controller = new AbortController()
-    const id = setTimeout(() => {
-        controller.abort()
-    }, timeout)
-    const response = await get(url, {
-        ...requestConfig,
-        signal: controller.signal,
-    })
-    clearTimeout(id)
-    return response
 }
 
 async function getFetch() {
