+++ sdk/nodejs/__tests__/eventQueue.test.ts
@@ -1,4 +1,4 @@
-jest.mock('../../js-cloud-server/src/request')
+jest.mock('../src/request')
 import { EventQueue, EventQueueOptions } from '../src/eventQueue'
 import { EventTypes } from '../src/eventQueue'
 import { BucketedUserConfig, DVCReporter, PublicProject } from '@devcycle/types'
@@ -10,8 +10,9 @@ import {
 } from '../src/bucketing'
 import { setPlatformDataJSON } from './utils/setPlatformData'
 import { Response } from 'cross-fetch'
-import { dvcDefaultLogger, publishEvents } from '@devcycle/js-cloud-server-sdk'
+import { dvcDefaultLogger } from '@devcycle/js-cloud-server-sdk'
 import { DVCPopulatedUserFromDevCycleUser } from '../src/models/populatedUserHelpers'
+import { publishEvents } from '../src/request'
 
 import testData from '@devcycle/bucketing-test-data/json-data/testData.json'
 const { config } = testData
